improve cart controller
